+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Simple Bayes Changelog
 
+## master
+
+- Fixed Bernoulli model's float overflow when dealing with large amount of data
+
 ## v0.9.0 [2016-08-04]
 
 - Added `top` option for restricting number of results